SearchState에 있는 로컬 모델 분리
